Move SCPRequestPipeline into package with all classes that mention it
Prelude to restricting it
