+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/WriteMemoryProcess.java
@@ -20,9 +20,9 @@ import static java.lang.Math.max;
 import static java.lang.Math.min;
 import static java.nio.ByteBuffer.allocate;
 import static org.apache.commons.io.IOUtils.buffer;
-import static uk.ac.manchester.spinnaker.connections.SCPRequestPipeline.SCP_RETRIES;
-import static uk.ac.manchester.spinnaker.connections.SCPRequestPipeline.SCP_TIMEOUT;
 import static uk.ac.manchester.spinnaker.messages.Constants.UDP_MESSAGE_MAX_SIZE;
+import static uk.ac.manchester.spinnaker.transceiver.SCPRequestPipeline.SCP_RETRIES;
+import static uk.ac.manchester.spinnaker.transceiver.SCPRequestPipeline.SCP_TIMEOUT;
 
 import java.io.File;
 import java.io.FileInputStream;
